Created endpoint: registration. Repaired some bugs

+++ backend/app/db/migrations/20230418110259-Users.js
@@ -1,44 +1,44 @@
-"use strict";
+'use strict';
 
-const { DataTypes } = require("sequelize");
+const { DataTypes } = require('sequelize');
 /** @type {import('sequelize-cli').Migration} */
 module.exports = {
-  async up(queryInterface, Sequelize) {
-    await queryInterface.createTable("users", {
-      id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-      },
-      name: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      email: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      password: {
-        type: DataTypes.STRING,
-        allowNull: false,
-      },
-      role_id: {
-        type: DataTypes.INTEGER,
-        allowNull: false,
-      },
-      createAt: {
-        type: DataTypes.DATE,
-        defaultValue: DataTypes.NOW,
-      },
-      ticket_id: {
-        type: DataTypes.INTEGER,
-        allowNull: true,
-      },
-    });
-  },
+    async up(queryInterface, Sequelize) {
+        await queryInterface.createTable('users', {
+            id: {
+                type: DataTypes.INTEGER,
+                allowNull: false,
+                autoIncrement: true,
+                primaryKey: true,
+            },
+            name: {
+                type: DataTypes.STRING,
+                allowNull: false,
+            },
+            email: {
+                type: DataTypes.STRING,
+                allowNull: false,
+            },
+            password: {
+                type: DataTypes.STRING,
+                allowNull: false,
+            },
+            role_id: {
+                type: DataTypes.INTEGER,
+                allowNull: false,
+            },
+            createAt: {
+                type: DataTypes.DATE,
+                defaultValue: DataTypes.NOW,
+            },
+            ticket_id: {
+                type: DataTypes.BIGINT,
+                allowNull: true,
+            },
+        });
+    },
 
-  async down(queryInterface, Sequelize) {
-    await queryInterface.dropTable("users");
-  },
+    async down(queryInterface, Sequelize) {
+        await queryInterface.dropTable('users');
+    },
 };
